Create README.md
This app will show you, in a google map,  the cafes loacated in London, UK that have WifFi That you can visit.
For that you need to register by filling the registration form.  Being a member allows you to add any cafe to the database , do updates to your account, delete a cafe if the info is wrong.